refactor(app): #373 Phase 1-3 use-terminal-tabs.ts に terminal tab を切り出し
App.tsx 2341 -> 2107 行 (-234)。terminal tab の state container と
自己完結ハンドラを src/renderer/src/lib/hooks/use-terminal-tabs.ts に集約。

移管:
- state: terminalTabs / activeTerminalTabId / nextTerminalIdRef /
  terminalActivityTimers / activeTerminalIds / tabCreateMenuOpen /
  pendingTeamClose / dragTabId / dragOverTabId / editingLabelTabId
- handler: addTerminalTab / closeTerminalTab / doCloseTab /
  restartTerminalTab / restartTerminal / markTerminalActivity
- 派生値: standaloneTabList / teamGroupList の useMemo
- 定数 / 純粋関数: MAX_TERMINALS / TERMINAL_WARN_THRESHOLD /
  getRoleDisplayLabel / TerminalTab interface
- effect: 初回タブ自動生成 (viewMode === 'ide' ガード付き) と
  activity timer の cleanup
- DnD: onDragStart / Over / Leave / Drop / End の 5 ハンドラを
  getDnDProps(tabId) factory にまとめて JSX 側でスプレッド可能に

Phase 1-1 / 1-2 の流儀に揃え、opts (viewMode / claudeReady /
projectRoot / showToast / teams / closeTeam) を optsRef.current で
毎 render 更新して内部 useCallback の deps を最小化。

doCloseTeam は teams setter / clearSpawnTimers / cleanupTeamMcp に
依存するため Phase 1-3 では App.tsx 残置。closeTeamRef ブリッジで
hook の opts.closeTeam に注入し循環を解消。teams state そのものも
Phase 1-4 (use-team-management) まで App.tsx で保持。

projectSwitchedRef.current の terminal リセット部分は hook が expose
する resetForProjectSwitch() に圧縮 (新規 Claude #1 を 1 個生成して
active に)。terminal リセットの責務は hook に閉じる。

terminalRefs (TerminalViewHandle Map) は <TerminalView> JSX が
App.tsx に残るため hook 化対象外。Phase 1-4 で再検討。

検証:
- npm run typecheck: green
- cargo check --manifest-path src-tauri/Cargo.toml: green
- cargo clippy --no-deps: 既存ベースライン 15 件のまま (新規警告ゼロ)

累計削減: -688 行 (#373 全体目標 -1995)。

Refs #373
